requirements.txt
discord.py~=1.6
Pillow~=8.0.1
aiohttp[speedups]~=3.6.3
aiofiles~=0.6.0
imgkit~=1.0.2
python-magic-win64~=0.4.13